Renamed all Pascal case (source) files to snake case
