+++ src/standard_modbus_functions.cpp
@@ -1,8 +1,8 @@
-#include "modbus_client_application.h"
+#include "standard_modbus_functions.h"
 #include <stdint.h>
 #include "function_code_parameters.h"
 #include "modbus_client.h"
-#include "MessagePriority.h"
+#include "message_priority.h"
 
 namespace DefaultModbusFunctions {
