--- src/types.h
@@ -1,25 +1,0 @@
-/*
- * @file types.h
- *
- * @brief Defines types used throughout Orcabrains600
- *
- *  Created on: Jan. 26, 2022
- *      Author: sjeffery
- */
-
-#ifndef TYPES_H_
-#define TYPES_H_
-
-#include <stdint.h>
-
-typedef uint8_t				u8;
-typedef int8_t				s8;
-typedef uint16_t			u16;
-typedef int16_t				s16;
-typedef uint32_t			u32;
-typedef int32_t				s32;
-typedef uint64_t			u64;
-typedef int64_t				s64;
-
-
-#endif /* TYPES_H_ */
