+++ tests/UnitAndIntegration/IrisSDKTests/modbus/TransactionConstructionTests.cpp
@@ -1,5 +1,5 @@
 #include "pch.h"
-#include "src/modbus_client_application.h"
+#include "src/standard_modbus_functions.h"
 #include "modbus/helpers/TestSerialInterface.h"
 #include "modbus/helpers/TestClock.h"
 #include "actuator.h"
